Update tests for rspec-puppet 2.0 + future parser
This patch updates raise_error and params_value to use `catalogue`
instead of `subject` since the contents of `subject` aren't evaluated
the way they used to be. Also, since the future parser considers empty
string to be a truthy value, and then interprets two incompatible
parameters to both be set if one is an empty string, it removes the
unset ipaddress parameter from one of the tests. Finally, it unpins
rspec-puppet.